requirements.txt
# Web爬虫相关
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
selenium>=4.15.0
scrapy>=2.11.0

# 数据处理
pandas>=2.1.0
numpy>=1.24.3
feedparser>=6.0.10

# 数据存储
sqlalchemy>=2.0.20
pymongo>=4.5.0
redis>=4.6.0

# 配置管理
pyyaml>=6.0.1
python-dotenv>=1.0.0

# 日志和监控
loguru>=0.7.2
schedule>=1.2.0

# HTTP代理和重试
requests-oauthlib>=1.3.1
retrying>=1.3.4

# 数据验证
pydantic>=2.4.0

# 工具库
click>=8.1.7
tqdm>=4.66.0
fake-useragent>=1.4.0
